[Fix] FollowService 예외 추가
ref: #28

+++ backend/src/main/java/site/bookmore/bookmore/users/service/FollowService.java
@@ -6,6 +6,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import site.bookmore.bookmore.common.exception.bad_request.FollowNotMeException;
+import site.bookmore.bookmore.common.exception.conflict.DuplicateFollowException;
+import site.bookmore.bookmore.common.exception.conflict.DuplicateUnfollowException;
 import site.bookmore.bookmore.common.exception.not_found.FollowNotFoundException;
 import site.bookmore.bookmore.common.exception.not_found.UserNotFoundException;
 import site.bookmore.bookmore.users.dto.FollowerResponse;
@@ -24,7 +26,7 @@ public class FollowService {
     private final UserRepository userRepository;
 
     public String following(Long id, String email) {
-        //Todo. 이미 팔로우한 경우 예외 처리.
+
         //나
         User user = userRepository.findByEmail(email)
                 .orElseThrow(UserNotFoundException::new);
@@ -38,10 +40,18 @@ public class FollowService {
             throw new FollowNotMeException();
         }
 
-        Follow follow = Follow.builder()
-                .following(targetUser)
-                .follower(user)
-                .build();
+        //이미 팔로우 한 경우(이전에 언팔도 한 적 없음)
+        followRepository.findByFollowerAndFollowingAndDeletedDatetimeIsNull(user, targetUser)
+                .ifPresent(follow -> {
+                    throw new DuplicateFollowException();
+                });
+
+        //이전에 언팔한 경우에 다시 팔로우 하는 경우
+        Follow follow = followRepository.findByFollowerAndFollowing(user, targetUser)
+                .orElse(Follow.builder()
+                        .following(targetUser)
+                        .follower(user)
+                        .build());
 
         follow.undelete();
 
@@ -64,25 +74,25 @@ public class FollowService {
         Follow targetFollow = followRepository.findByFollowerAndFollowing(user, targetUser)
                 .orElseThrow(FollowNotFoundException::new);
 
+        //언팔로우를 이미 한 경우
+        if (targetFollow.getDeletedDatetime() != null) {
+            throw new DuplicateUnfollowException();
+        }
+
         targetFollow.delete();
 
         return String.format("%s 님을 언팔로우 하셨습니다.", id);
     }
 
     public Page<FollowingResponse> findAllFollowing(Long id, Pageable pageable) {
-        // Todo. 소프트 삭제 제외하고 조회.
-        userRepository.findById(id).orElseThrow(UserNotFoundException::new);
 
-        return followRepository.findAll(pageable).map(follow -> new FollowingResponse(follow));
-
+        User user = userRepository.findById(id).orElseThrow(UserNotFoundException::new);
+        return followRepository.findByFollowerAndDeletedDatetimeIsNull(pageable, user).map(FollowingResponse::new);
     }
 
     public Page<FollowerResponse> findAllFollower(Long id, Pageable pageable) {
 
-        userRepository.findById(id).orElseThrow(UserNotFoundException::new);
-
-        Page<Follow> follows = followRepository.findAll(pageable);
-
-        return followRepository.findAll(pageable).map(follow -> new FollowerResponse(follow));
+        User user = userRepository.findById(id).orElseThrow(UserNotFoundException::new);
+        return followRepository.findByFollowingAndDeletedDatetimeIsNull(pageable, user).map(FollowerResponse::new);
     }
 }
